requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.0
